Extract upload logic into its own class.
